Refactor namespaces and update README for clarity; enhance command options
