+++ demo/Program.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Blazor.Components;
+using Blazor.CommandLine;
 
 var builder = WebApplication.CreateBuilder(args);
