+++ demo/_Imports.razor
@@ -5,8 +5,8 @@
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.JSInterop
-@using Blazor.Components
-@using Blazor.Components.CommandLine
+@using Blazor.CommandLine
+@using Blazor.CommandLine.Command
 @using BlazorConsoleDemo
 @using BlazorConsoleDemo.Shared
