+++ src/Blazor.CommandLine/BlazorCommandLine.razor
@@ -1,7 +1,7 @@
 @using Microsoft.AspNetCore.Components.Forms
-@using Blazor.Components.CommandLine;
+@using Blazor.CommandLine;
 
-@inherits Blazor.Components.BlazorCommandLineComponent
+@inherits Blazor.CommandLine.BlazorCommandLineComponent
 
 <div id="console">
     <div id="header">
