+++ src/Blazor.CommandLine/Command/RunningCommand.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Blazor.Components.CommandLine;
+namespace Blazor.CommandLine.Command;
 
 public class RunningCommand : IRunningCommand
 {
