+++ src/Blazor.CommandLine/ServiceCollectionExtensions.cs
@@ -1,25 +1,24 @@
-﻿using Blazor.Components.CommandLine;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.DependencyInjection;
 using System;
+using Blazor.CommandLine.Command;
 
-namespace Blazor.Components
+namespace Blazor.CommandLine;
+
+public static class ServiceCollectionExtensions
 {
-    public static class ServiceCollectionExtensions
+    public static IServiceCollection AddCommandLine(this IServiceCollection serviceCollection,
+        Action<CommandLineOptions> options = null)
     {
-        public static IServiceCollection AddCommandLine(this IServiceCollection serviceCollection,Action<CommandLineOptions> options=null)
-        {
+        serviceCollection.AddTransient<IRunningCommand, RunningCommand>();
+        var consoleOptions = new CommandLineOptions();
 
-            serviceCollection.AddTransient<IRunningCommand, RunningCommand>();
-            var consoleOptions = new CommandLineOptions();
+        options?.Invoke(consoleOptions);
 
-            options?.Invoke(consoleOptions);
+        return serviceCollection;
+    }
 
-            return serviceCollection;
-        }
-
-        public class CommandLineOptions
-        {
-            public bool UseDefaultServices { get; set; } = true;
-        }
+    public class CommandLineOptions
+    {
+        public bool UseDefaultServices { get; set; } = true;
     }
 }
